--- src/stories/Button.stories.js
@@ -1,9 +1,0 @@
-import React from 'react'
-import Button from '../Button'
-
-export default {
-  title: 'Button',
-  component: Button,
-}
-
-export const DefaultButton = () => <Button>Default Button</Button>
