+++ src/stories/StyledButton.stories.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import StyledButton from '../StyledButton'
+
+export default {
+  title: 'StyledButton',
+  component: StyledButton,
+}
+
+export const DefaultButton = () => <StyledButton>Default Button</StyledButton>
